Changed most physical dimenions to decimal type. Using Doubles caused small rounding errors when summing and comparing values.

+++ PBJJ/Core/NewProfileViewModel.cs
@@ -2,7 +2,7 @@ namespace PBJJ.Core {
     public class NewProfileViewModel {
         public string Type { get; set; }
         public string Name { get; set; }
-        public double FingerWidth { get; set; }
-        public double OverallWidth { get; set; }
+        public decimal FingerWidth { get; set; }
+        public decimal OverallWidth { get; set; }
     }
 }
