+++ PBJJ/Core/ProfileGenerator.cs
@@ -8,7 +8,7 @@ namespace PBJJ.Core
 {
     public static class ProfileGenerator
     {
-        public static JointProfile GenerateStandardProfile(double fingerWidthInches, double overallWidthInches)
+        public static JointProfile GenerateStandardProfile(decimal fingerWidthInches, decimal overallWidthInches)
         {
             string name = $"Standard {fingerWidthInches:N3}in fingers over {overallWidthInches}in";
 
@@ -16,7 +16,7 @@ namespace PBJJ.Core
 
             JointProfileElement.JointProfileElementType currentElementType = JointProfileElement.JointProfileElementType.Finger;
 
-            double currentWidth = 0;
+            decimal currentWidth = 0;
             do
             {
                 profile.Elements.Add(new JointProfileElement(currentElementType, fingerWidthInches));
